ldr and soil value float to int in db and websocket fix

+++ iot_app/models.py
@@ -9,8 +9,8 @@ class PlantData(Base):
     date = Column(String, index=True)
     temperature = Column(Integer)
     humidity = Column(Integer)
-    lightval = Column(Float)
-    moisture = Column(Float)
+    lightval = Column(Integer)
+    moisture = Column(Integer)
 
 
 class EmailData(Base):
